requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
google-generativeai
openai
pydantic
requests
langchain
langchain-google-genai
langchain-openai
